chore (dedup): reorganize deduplication and merge
_deduplicate_and_merge becomes a SearchResult method
creation of utilitaries in utils/dedup.py
adding title_year_key as a method of Paper class
Modification of tests to match the new organization
Duplication of deduplication tests from search_runner into
search_result.
Test run ok Previous behavior is unchanged
