+++ findpapers/runners/search_runner.py
@@ -22,33 +22,6 @@ from findpapers.utils.progress import make_progress_bar
 
 logger = logging.getLogger(__name__)
 
-_PREPRINT_DOI_PREFIXES: frozenset[str] = frozenset(
-    {
-        "10.48550/arxiv.",  # arXiv
-        "10.1101/",  # bioRxiv / medRxiv
-        "10.2139/ssrn.",  # SSRN
-        "10.5281/zenodo.",  # Zenodo
-        "10.20944/preprints",  # Preprints.org
-    }
-)
-
-
-def _is_preprint_doi(doi: str) -> bool:
-    """Return ``True`` when *doi* belongs to a preprint server.
-
-    Parameters
-    ----------
-    doi : str
-        DOI string (without ``https://doi.org/`` prefix).
-
-    Returns
-    -------
-    bool
-        ``True`` for known preprint-server DOI prefixes.
-    """
-    lowered = doi.strip().lower()
-    return any(lowered.startswith(prefix) for prefix in _PREPRINT_DOI_PREFIXES)
-
 
 class SearchRunner(DiscoveryRunner):
     """Public API entry point for running academic paper searches.
@@ -288,31 +261,41 @@ class SearchRunner(DiscoveryRunner):
             for searcher in self._searchers:
                 searcher.close()
 
+        self._search = SearchResult(
+            query=self._query_string,
+            max_papers_per_database=self._max_papers_per_database,
+            processed_at=datetime.now(UTC),
+            databases=[s.name for s in self._searchers],
+            papers=list(self._results),
+            runtime_seconds=None,
+            runtime_seconds_per_database=db_runtimes or None,
+            since=self._since,
+            until=self._until,
+            failed_databases=failed_databases or None,
+        )
         before_dedupe = len(self._results)
-        self._deduplicate_and_merge(metrics)
+        self._search._deduplicate_and_merge(metrics)
         if verbose:
-            merged = before_dedupe - len(self._results)
+            merged = before_dedupe - len(self._search.papers)
             logger.info(
                 "Dedupe: %d -> %d papers (%d merged)",
                 before_dedupe,
                 len(self._results),
                 merged,
             )
-
         # Apply post-fetch date filters.  Connectors that only support
         # year-level date filtering may return papers outside the requested
         # range; _matches_filters enforces precise boundaries.
         if self._since is not None or self._until is not None:
-            before_filter = len(self._results)
-            self._results = [p for p in self._results if self._matches_filters(p)]
+            before_filter = len(self._search.papers)
+            self._search.papers = [p for p in self._search.papers if self._matches_filters(p)]
             if verbose:
                 logger.info(
                     "Post-fetch filter: %d -> %d papers (%d removed)",
                     before_filter,
-                    len(self._results),
-                    before_filter - len(self._results),
+                    len(self._search.papers),
+                    before_filter - len(self._search.papers),
                 )
-
         # Enrich the filtered papers via per-paper get() lookups.
         # enrichment_databases=None  → enrich with all available databases.
         # enrichment_databases=[]    → skip enrichment entirely.
@@ -320,13 +303,12 @@ class SearchRunner(DiscoveryRunner):
             isinstance(self._enrichment_databases, list) and len(self._enrichment_databases) == 0
         ):
             super()._enrich_papers(
-                self._results,
+                self._search.papers,
                 verbose,
                 show_progress=show_progress,
                 num_workers=self._num_workers,
             )
-
-        metrics["total_papers"] = len(self._results)
+        metrics["total_papers"] = len(self._search.papers)
         metrics["runtime_in_seconds"] = perf_counter() - start
         self._metrics = metrics
 
@@ -337,19 +319,9 @@ class SearchRunner(DiscoveryRunner):
             for searcher in self._searchers:
                 count = int(metrics.get(f"total_papers_from_{searcher.name}", 0))
                 logger.info("  %s: %d papers", searcher.name, count)
-
-        self._search = SearchResult(
-            query=self._query_string,
-            max_papers_per_database=self._max_papers_per_database,
-            processed_at=datetime.now(UTC),
-            databases=[s.name for s in self._searchers],
-            papers=list(self._results),
-            runtime_seconds=self._metrics.get("runtime_in_seconds"),
-            runtime_seconds_per_database=db_runtimes or None,
-            since=self._since,
-            until=self._until,
-            failed_databases=failed_databases or None,
-        )
+        # Update of search attributes
+        self._search.runtime_seconds = self._metrics.get("runtime_in_seconds")
+        self._search.processed_at = datetime.now(UTC)
         _root_logger.setLevel(_saved_log_level)
         return self._search
 
@@ -576,170 +548,3 @@ class SearchRunner(DiscoveryRunner):
                 pbar.close()
 
         return failed, db_runtimes
-
-    def _deduplicate_and_merge(self, metrics: dict[str, int | float]) -> None:
-        """Collapse duplicate papers in two passes.
-
-        **Pass 1** groups papers by their primary key (DOI when available,
-        otherwise a normalised ``title|year`` string).  This resolves exact
-        duplicates found within the same database or between databases that
-        share the same DOI.
-
-        **Pass 2** groups the results of pass 1 by normalised title and
-        merges entries whose publication years are *compatible* — i.e. they
-        share the same year, at least one has no year (incomplete metadata),
-        or at least one entry carries a preprint DOI and their years differ by
-        at most one (to handle both the case of the same preprint deposited to
-        two servers across the Dec/Jan calendar boundary and the common
-        preprint-to-published transition, e.g. Zenodo 2026 + book chapter
-        2025).  Papers with the same title and *different known years* where
-        neither is from a preprint server are intentionally kept separate.
-
-        This correctly handles the common cross-database case where the same
-        work is indexed with different DOIs (e.g. an arXiv preprint DOI vs a
-        publisher DOI) and one of the database records lacks a publication
-        date.
-
-        When two papers are deemed duplicates their data is merged using
-        :meth:`~findpapers.core.paper.Paper.merge` (most-complete strategy).
-
-        Parameters
-        ----------
-        metrics : dict[str, int | float]
-            Metrics dict (currently unused; reserved for future statistics).
-
-        Returns
-        -------
-        None
-        """
-        # Pass 1: primary key dedup (DOI > title|year > title).
-        pass1: dict[str, Paper] = {}
-        for paper in self._results:
-            key = self._dedupe_key(paper)
-            if key in pass1:
-                pass1[key].merge(paper)
-            else:
-                pass1[key] = paper
-
-        # Pass 2: title-based dedup that handles missing year metadata.
-        # Group survivors from pass 1 by normalised title, then within each
-        # title group greedily merge papers whose years are compatible.
-        by_title: dict[str, list[Paper]] = {}
-        untitled: list[Paper] = []
-        for paper in pass1.values():
-            norm_title = paper.title.strip().lower() if paper.title else ""
-            if norm_title:
-                by_title.setdefault(norm_title, []).append(paper)
-            else:
-                untitled.append(paper)
-
-        result: list[Paper] = list(untitled)
-        for candidates in by_title.values():
-            # Greedily merge into the first compatible representative.
-            groups: list[Paper] = []
-            for paper in candidates:
-                paper_year = getattr(paper.publication_date, "year", None)
-                merged_into: Paper | None = None
-                for representative in groups:
-                    rep_year = getattr(representative.publication_date, "year", None)
-                    if _are_years_compatible(rep_year, paper_year, representative.doi, paper.doi):
-                        merged_into = representative
-                        break
-                if merged_into is not None:
-                    merged_into.merge(paper)
-                else:
-                    groups.append(paper)
-            result.extend(groups)
-
-        self._results = result
-
-    def _dedupe_key(self, paper: Paper) -> str:
-        """Build a stable primary deduplication key for a paper.
-
-        Uses the DOI when available; otherwise falls back to a normalised
-        ``title|year`` combination.
-
-        Parameters
-        ----------
-        paper : Paper
-            Paper to key.
-
-        Returns
-        -------
-        str
-            Dedupe key string.
-        """
-        if paper.doi:
-            return f"doi:{str(paper.doi).strip().lower()}"
-        return self._title_year_key(paper)
-
-    def _title_year_key(self, paper: Paper) -> str:
-        """Build a normalised ``title|year`` deduplication key for a paper.
-
-        Used as the pass-1 fallback dedup key when no DOI is available.
-        Two records without a DOI but with the same title and the same year
-        are considered identical and are merged in pass 1.
-
-        Parameters
-        ----------
-        paper : Paper
-            Paper to key.
-
-        Returns
-        -------
-        str
-            Title/year key string.
-        """
-        title = paper.title
-        year = getattr(paper.publication_date, "year", None)
-        if title and year:
-            return f"title:{str(title).strip().lower()}|year:{year}"
-        if title:
-            return f"title:{str(title).strip().lower()}"
-        return f"object:{id(paper)}"
-
-
-def _are_years_compatible(
-    year_a: int | None,
-    year_b: int | None,
-    doi_a: str | None,
-    doi_b: str | None,
-) -> bool:
-    """Return whether two publication years are compatible for merging.
-
-    Two papers are considered year-compatible when:
-
-    * Either year is unknown (``None``), or
-    * Both years are identical, or
-    * Their years differ by exactly 1 **and** at least one DOI belongs to a
-      preprint server (covers preprint-to-published transitions across the
-      Dec/Jan boundary).
-
-    Parameters
-    ----------
-    year_a : int | None
-        Publication year of the first paper.
-    year_b : int | None
-        Publication year of the second paper.
-    doi_a : str | None
-        DOI of the first paper.
-    doi_b : str | None
-        DOI of the second paper.
-
-    Returns
-    -------
-    bool
-        ``True`` when the years are compatible for merging.
-    """
-    if year_a is None or year_b is None or year_a == year_b:
-        return True
-
-    # Adjacent-year preprint check.
-    raw_doi_a = doi_a or ""
-    raw_doi_b = doi_b or ""
-    return (
-        abs(year_a - year_b) == 1
-        and bool(raw_doi_a)
-        and bool(raw_doi_b)
-        and (_is_preprint_doi(raw_doi_a) or _is_preprint_doi(raw_doi_b))
-    )
